refactor trace writer externalize traversal
